refactor: extract logic to composables, organize structure and routing guards
- Moved validation and business logic out of components into dedicated composables
- Grouped composables by domain into separate directories for better structure
- Moved route access checks (e.g. form validation for /result) into middleware
- Improved code readability and maintainability by applying separation of concerns
