+++ components/AppFooter.vue
@@ -1,31 +1,28 @@
-<script setup>
-</script>
-
 <template>
   <footer
-      class="footer border-t-5 border-(--color-primary) flex items-center justify-between pl-5 pr-5 h-20 md:h-25 bg-(--background-toolbar)"
+    class="footer border-t-5 border-(--color-primary) flex items-center justify-between pl-5 pr-5 h-20 md:h-25 bg-(--background-toolbar)"
   >
     <div class="max-w-7xl w-full m-auto flex justify-between">
       <div class="flex items-center">
         <a href="https://kknights.com/" target="_blank">
           <svg
-              class="fill-(--color-primary) h-7.5"
-              viewBox="0 0 30.1 32.01"
-              xmlns="http://www.w3.org/2000/svg"
+            class="fill-(--color-primary) h-7.5"
+            viewBox="0 0 30.1 32.01"
+            xmlns="http://www.w3.org/2000/svg"
           >
             <g id="Layer_2" data-name="Layer 2">
               <g id="Layer_1-2" data-name="Layer 1">
                 <path
-                    class="stroke"
-                    d="M25.28,4.47A11.88,11.88,0,0,0,22.68,6a.2.2,0,0,0,0,.35,7.51,7.51,0,0,1,2.4,2.57c.33.59.79.58,1.07,0a5.92,5.92,0,0,0-.06-4.08A.59.59,0,0,0,25.28,4.47Z"
+                  class="stroke"
+                  d="M25.28,4.47A11.88,11.88,0,0,0,22.68,6a.2.2,0,0,0,0,.35,7.51,7.51,0,0,1,2.4,2.57c.33.59.79.58,1.07,0a5.92,5.92,0,0,0-.06-4.08A.59.59,0,0,0,25.28,4.47Z"
                 />
                 <path
-                    class="stroke"
-                    d="M29.26,1.23A1.9,1.9,0,0,0,27.09,0a24.88,24.88,0,0,0-7.25,2.64,1.38,1.38,0,0,1-1,.18l-.58-.12a12.26,12.26,0,0,0-3.23-.37,12.26,12.26,0,0,0-3.23.37l-.58.12a1.38,1.38,0,0,1-1-.18A24.88,24.88,0,0,0,3,0,1.9,1.9,0,0,0,.84,1.23C-.31,4.4-.18,11.39.66,15.67l0,.21.21,11.5a3.47,3.47,0,0,0,2.32,3.21A22.86,22.86,0,0,0,11.49,32h7.12a22.86,22.86,0,0,0,8.28-1.42,3.47,3.47,0,0,0,2.32-3.21l.21-11.5,0-.21C30.28,11.39,30.41,4.4,29.26,1.23ZM27.48,15.28l-.06.56-.21,11.51a1.49,1.49,0,0,1-1,1.36,20.64,20.64,0,0,1-6.8,1.3.79.79,0,0,1-.79-.79c0-1,0-2.83,0-3.83a1,1,0,0,1,.7-1L23.68,23a1.63,1.63,0,0,0,1.15-1.5L25,15.7a1.61,1.61,0,0,0-1.85-1.65q-4,.4-8.12.41h0q-4.06,0-8.12-.41A1.61,1.61,0,0,0,5.07,15.7l.2,5.76A1.63,1.63,0,0,0,6.42,23l4.36,1.46a1,1,0,0,1,.7,1v3.83a.79.79,0,0,1-.79.79,20.58,20.58,0,0,1-6.8-1.3,1.47,1.47,0,0,1-1-1.36L2.68,15.84l-.06-.56A36.65,36.65,0,0,1,2.42,3a1,1,0,0,1,1.25-.77A21.78,21.78,0,0,1,9.24,4.4a3.49,3.49,0,0,0,2.39.43l.64-.14A10.05,10.05,0,0,1,15,4.37h0a10.05,10.05,0,0,1,2.77.32l.64.14a3.49,3.49,0,0,0,2.39-.43,21.78,21.78,0,0,1,5.57-2.16A1,1,0,0,1,27.68,3,36.65,36.65,0,0,1,27.48,15.28ZM8.71,20H7.39a.52.52,0,0,1-.52-.52V18.28a.52.52,0,0,1,.52-.52H8.71a.37.37,0,0,0,.37-.37V16.07a.52.52,0,0,1,.52-.52h1.16a.52.52,0,0,1,.52.52v1.32a.36.36,0,0,0,.37.37H13a.52.52,0,0,1,.52.52v1.16A.52.52,0,0,1,13,20H11.63a.38.38,0,0,0-.37.37v1.32a.52.52,0,0,1-.52.52H9.6a.52.52,0,0,1-.52-.52V20.33A.38.38,0,0,0,8.71,20Zm12.18-1.17A1.18,1.18,0,1,1,22.07,20,1.18,1.18,0,0,1,20.89,18.79Zm.3,2.06A1.18,1.18,0,1,1,20,19.67,1.18,1.18,0,0,1,21.19,20.85ZM20,17.91a1.18,1.18,0,1,1,1.18-1.18A1.18,1.18,0,0,1,20,17.91Zm-.88.88A1.18,1.18,0,1,1,18,17.61,1.18,1.18,0,0,1,19.13,18.79Zm-2.4,5.44L15.6,25.4h0a.72.72,0,0,1-1,0l-1.13-1.12a.77.77,0,0,1-.22-.53A.73.73,0,0,1,14,23h2.26a.73.73,0,0,1,.52,1.2Z"
+                  class="stroke"
+                  d="M29.26,1.23A1.9,1.9,0,0,0,27.09,0a24.88,24.88,0,0,0-7.25,2.64,1.38,1.38,0,0,1-1,.18l-.58-.12a12.26,12.26,0,0,0-3.23-.37,12.26,12.26,0,0,0-3.23.37l-.58.12a1.38,1.38,0,0,1-1-.18A24.88,24.88,0,0,0,3,0,1.9,1.9,0,0,0,.84,1.23C-.31,4.4-.18,11.39.66,15.67l0,.21.21,11.5a3.47,3.47,0,0,0,2.32,3.21A22.86,22.86,0,0,0,11.49,32h7.12a22.86,22.86,0,0,0,8.28-1.42,3.47,3.47,0,0,0,2.32-3.21l.21-11.5,0-.21C30.28,11.39,30.41,4.4,29.26,1.23ZM27.48,15.28l-.06.56-.21,11.51a1.49,1.49,0,0,1-1,1.36,20.64,20.64,0,0,1-6.8,1.3.79.79,0,0,1-.79-.79c0-1,0-2.83,0-3.83a1,1,0,0,1,.7-1L23.68,23a1.63,1.63,0,0,0,1.15-1.5L25,15.7a1.61,1.61,0,0,0-1.85-1.65q-4,.4-8.12.41h0q-4.06,0-8.12-.41A1.61,1.61,0,0,0,5.07,15.7l.2,5.76A1.63,1.63,0,0,0,6.42,23l4.36,1.46a1,1,0,0,1,.7,1v3.83a.79.79,0,0,1-.79.79,20.58,20.58,0,0,1-6.8-1.3,1.47,1.47,0,0,1-1-1.36L2.68,15.84l-.06-.56A36.65,36.65,0,0,1,2.42,3a1,1,0,0,1,1.25-.77A21.78,21.78,0,0,1,9.24,4.4a3.49,3.49,0,0,0,2.39.43l.64-.14A10.05,10.05,0,0,1,15,4.37h0a10.05,10.05,0,0,1,2.77.32l.64.14a3.49,3.49,0,0,0,2.39-.43,21.78,21.78,0,0,1,5.57-2.16A1,1,0,0,1,27.68,3,36.65,36.65,0,0,1,27.48,15.28ZM8.71,20H7.39a.52.52,0,0,1-.52-.52V18.28a.52.52,0,0,1,.52-.52H8.71a.37.37,0,0,0,.37-.37V16.07a.52.52,0,0,1,.52-.52h1.16a.52.52,0,0,1,.52.52v1.32a.36.36,0,0,0,.37.37H13a.52.52,0,0,1,.52.52v1.16A.52.52,0,0,1,13,20H11.63a.38.38,0,0,0-.37.37v1.32a.52.52,0,0,1-.52.52H9.6a.52.52,0,0,1-.52-.52V20.33A.38.38,0,0,0,8.71,20Zm12.18-1.17A1.18,1.18,0,1,1,22.07,20,1.18,1.18,0,0,1,20.89,18.79Zm.3,2.06A1.18,1.18,0,1,1,20,19.67,1.18,1.18,0,0,1,21.19,20.85ZM20,17.91a1.18,1.18,0,1,1,1.18-1.18A1.18,1.18,0,0,1,20,17.91Zm-.88.88A1.18,1.18,0,1,1,18,17.61,1.18,1.18,0,0,1,19.13,18.79Zm-2.4,5.44L15.6,25.4h0a.72.72,0,0,1-1,0l-1.13-1.12a.77.77,0,0,1-.22-.53A.73.73,0,0,1,14,23h2.26a.73.73,0,0,1,.52,1.2Z"
                 />
                 <path
-                    class="stroke"
-                    d="M4.81,4.47A.58.58,0,0,0,4,4.83,5.86,5.86,0,0,0,4,8.91c.28.61.74.62,1.07,0A7.51,7.51,0,0,1,7.43,6.37a.2.2,0,0,0,0-.35A11.67,11.67,0,0,0,4.81,4.47Z"
+                  class="stroke"
+                  d="M4.81,4.47A.58.58,0,0,0,4,4.83,5.86,5.86,0,0,0,4,8.91c.28.61.74.62,1.07,0A7.51,7.51,0,0,1,7.43,6.37a.2.2,0,0,0,0-.35A11.67,11.67,0,0,0,4.81,4.47Z"
                 />
               </g>
             </g>
@@ -34,15 +31,12 @@
       </div>
       <div class="flex items-center text-(--text-title)">
         <a
-            class="cursor-pointer inline-flex justify-center rounded-md border mr-4 border-transparent bg-(--button-donation) px-2 py-1 text-sm font-medium text-(--background-toolbar) opacity-80 hover:opacity-100 hover:text-(--text-reverted) focus:outline-none focus-visible:ring-2 focus-visible:ring-(color-primary) focus-visible:ring-offset-2"
-            href="https://yoomoney.ru/to/41001420208384"
-            target="_blank"
-        >На кофе :3</a
+          class="cursor-pointer inline-flex justify-center rounded-md border mr-4 border-transparent bg-(--button-donation) px-2 py-1 text-sm font-medium text-(--background-toolbar) opacity-80 hover:opacity-100 hover:text-(--text-reverted) focus:outline-none focus-visible:ring-2 focus-visible:ring-(color-primary) focus-visible:ring-offset-2"
+          href="https://yoomoney.ru/to/41001420208384"
+          target="_blank"
+          >На кофе :3</a
         >
-        <a
-            href="https://github.com/konstantinstukov/kkluck"
-            target="_blank"
-        >
+        <a href="https://github.com/konstantinstukov/kkluck" target="_blank">
           dev by KS
         </a>
       </div>
