+++ nuxt.config.ts
@@ -10,7 +10,7 @@ export default defineNuxtConfig({
     css: {
       preprocessorOptions: {
         scss: {
-          additionalData: '@import "@/assets/scss/_variables.scss";',
+          additionalData: '@use "@/assets/scss/_variables.scss" as *;',
         },
       },
     },
