+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kkluck",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "private": true,
   "type": "module",
   "packageManager": "bun@1.2.9",
